backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
google-generativeai==0.3.2
pyttsx3==2.90
pyobjc-framework-Cocoa==10.0
speechrecognition==3.10.0
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow==2.13.0
